requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
flake8==6.0.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.14
